+++ lib/react/jsx.rb
@@ -1,4 +1,5 @@
 require 'execjs'
+require 'react/jsx/processor'
 require 'react/jsx/template'
 require 'react/jsx/jsx_transformer'
 require 'react/jsx/babel_transformer'
@@ -9,7 +10,7 @@ module React
     DEFAULT_TRANSFORMER = BabelTransformer
     mattr_accessor :transform_options, :transformer_class, :transformer
 
-    # You can assign `React::JSX.transformer_class = `
+    # You can assign `config.react.jsx_transformer_class = `
     # to provide your own transformer. It must implement:
     # - #initialize(options)
     # - #transform(code) => new code