+++ lib/react/rails/railtie.rb
@@ -4,7 +4,6 @@ module React
   module Rails
     class Railtie < ::Rails::Railtie
       config.react = ActiveSupport::OrderedOptions.new
-
       # Sensible defaults. Can be overridden in application.rb
       config.react.variant = (::Rails.env.production? ? :production : :development)
       config.react.addons = false
@@ -15,6 +14,8 @@ module React
       config.react.server_renderer_timeout    = 20  # seconds
       config.react.server_renderer            = nil # defaults to SprocketsRenderer
       config.react.server_renderer_options    = {}  # SprocketsRenderer provides defaults
+      # View helper implementation:
+      config.react.view_helper_implementation = nil # Defaults to ComponentMount
 
       # Watch .jsx files for changes in dev, so we can reload the JS VMs with the new JS code.
       initializer "react_rails.add_watchable_files", group: :all do |app|
@@ -23,12 +24,29 @@ module React
 
       # Include the react-rails view helper lazily
       initializer "react_rails.setup_view_helpers", group: :all do |app|
+
         app.config.react.jsx_transformer_class ||= React::JSX::DEFAULT_TRANSFORMER
         React::JSX.transformer_class = app.config.react.jsx_transformer_class
         React::JSX.transform_options = app.config.react.jsx_transform_options
 
+        app.config.react.view_helper_implementation ||= React::Rails::ComponentMount
+        React::Rails::ViewHelper.helper_implementation_class = app.config.react.view_helper_implementation
+
+        ActiveSupport.on_load(:action_controller) do
+          include ::React::Rails::ControllerLifecycle
+        end
+
         ActiveSupport.on_load(:action_view) do
           include ::React::Rails::ViewHelper
+        end
+      end
+
+      initializer "react_rails.add_component_renderer", group: :all do |app|
+        ActionController::Renderers.add :component do |component_name, options|
+          renderer = ::React::Rails::ControllerRenderer.new(controller: self)
+          html = renderer.call(component_name, options)
+          render_options = options.merge(inline: html)
+          render(render_options)
         end
       end
 
@@ -43,7 +61,7 @@ module React
 
       end
 
-      config.before_initialize do |app|
+      initializer "react_rails.set_variant", after: :engines_blank_point, group: :all do |app|
         asset_variant = React::Rails::AssetVariant.new({
           variant: app.config.react.variant,
           addons: app.config.react.addons,