+++ lib/react/rails/view_helper.rb
@@ -1,28 +1,24 @@
 module React
   module Rails
     module ViewHelper
-      # Render a UJS-type HTML tag annotated with data attributes, which
-      # are used by react_ujs to actually instantiate the React component
-      # on the client.
-      def react_component(name, props = {}, options = {}, &block)
-        options = {:tag => options} if options.is_a?(Symbol)
+      # This class will be used for inserting tags into HTML.
+      # It should implement:
+      #   - #setup(controller_instance)
+      #   - #teardown(controller_instance)
+      #   - #react_component(name, props, options &block)
+      # The default is {React::Rails::ComponentMount}
+      mattr_accessor :helper_implementation_class
 
-        prerender_options = options[:prerender]
-        if prerender_options
-          block = Proc.new{ concat React::ServerRendering.render(name, props, prerender_options) }
-        end
-
-        html_options = options.reverse_merge(:data => {})
-        html_options[:data].tap do |data|
-          data[:react_class] = name
-          data[:react_props] = (props.is_a?(String) ? props : props.to_json)
-        end
-        html_tag = html_options[:tag] || :div
-
-        # remove internally used properties so they aren't rendered to DOM
-        html_options.except!(:tag, :prerender)
-
-        content_tag(html_tag, '', html_options, &block)
+      # Render a React component into the view
+      # using the {helper_implementation_class}
+      #
+      # If called during a Rails controller-managed request, use the instance
+      # created by the controller.
+      #
+      # Otherwise, make a new instance.
+      def react_component(*args, &block)
+        helper_obj = @__react_component_helper ||= helper_implementation_class.new
+        helper_obj.react_component(*args, &block)
       end
     end
   end