+++ test/dummy/app/controllers/server_controller.rb
@@ -14,4 +14,13 @@ class ServerController < ApplicationController
     React::ServerRendering.reset_pool
     @todos = %w{todo1 todo2 todo3}
   end
+
+  def inline_component
+    render component: 'TodoList',
+           props: { todos: ['Render this inline'] },
+           tag: 'span',
+           class: 'custom-class',
+           id: 'custom-id',
+           data: { remote: true }
+  end
 end